Update store methods for asynchronous operations (#1)
Since we are using IndexedDB, we are required to use asynchronous
methods.

The widespread use of these asynchronous methods throughout the
application may lead to increased complexity. One possible solution is
to load the asynchronous logic into memory partially.

There are two main directions we are considering:

A. Open up the asynchronous API of Storage to the maximum extent.

This approach may offer performance benefits but could increase the
complexity of the application. Asynchronous code can be more
challenging to debug and can lead to more complex exception handling.

B. Restrict API of Storage and only use asynchronous processing in
certain situations.

This approach may reduce the complexity of the application, but it may
impact performance. Asynchronous processing may be necessary for
scenarios involving large data or complex operations.